# QL_FOODDELI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QL_FOODDELI API (Dapper & SignalR)",
    "description": "Hệ thống backend API quản lý giao đồ ăn nhanh QL_FOODDELI - Sprint 5",
    "version": "v1"
  },
  "paths": {
...
    "/api/GiaoHang": {
      "get": {
        "tags": [
          "GiaoHang"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "GiaoHang"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GiaoHang"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GiaoHang"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GiaoHang"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/GiaoHang/{id}": {
      "get": {
        "tags": [
          "GiaoHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/GioHang/them-mon": {
      "post": {
        "tags": [
          "GioHang"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemMonRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemMonRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ThemMonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/GioHang/cap-nhat-so-luong": {
      "put": {
        "tags": [
          "GioHang"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatSoLuongRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatSoLuongRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatSoLuongRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/GioHang/xoa-mon/{id}": {
      "delete": {
        "tags": [
          "GioHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CapNhatSoLuongRequest": {
        "type": "object",
        "properties": {
          "maChiTietGioHang": {
            "type": "string",
            "nullable": true
          },
          "soLuong": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "GiaoHang": {
        "type": "object",
        "properties": {
          "maGiaoHang": {
            "type": "string",
            "nullable": true
          },
          "maDonHang": {
            "type": "string",
            "nullable": true
          },
          "maNguoiGiaoHang": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "thoiGianNhan": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "thoiGianGiao": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ThemMonRequest": {
        "type": "object",
        "properties": {
          "maMonAn": {
            "type": "string",
            "nullable": true
          },
          "soLuong": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Dán mã JWT Token thu được sau khi Đăng nhập vào ô này (KHÔNG gõ chữ Bearer):",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}